backend/requirements.txt
# DD-Analyst Backend Dependencies
# Pinned versions for reproducible builds

# --- Web Framework & Server ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1
python-multipart==0.0.20

# --- Database ---
asyncpg==0.30.0

# --- LLM Inference ---
anthropic>=0.84.0
openai>=1.40.0

# --- Vector Store ---
chromadb==0.5.23

# --- Embeddings & Reranking ---
sentence-transformers==3.3.1

# --- Document Parsers ---
PyMuPDF==1.25.1
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
extract-msg==0.52.0
python-magic-bin==0.4.14
chardet==5.2.0
striprtf==0.0.26
lxml==5.3.0

# --- OCR ---
# paddleocr==2.9.1  # Install separately: requires paddlepaddle
# pytesseract==0.3.13

# --- Image Processing ---
Pillow==11.1.0
# opencv-python-headless==4.10.0.84  # Install separately if OCR needed

# --- Text Processing ---
tiktoken==0.8.0

# --- Report Generation ---
reportlab==4.2.5
XlsxWriter==3.2.0

# --- Security ---
cryptography==44.0.0
argon2-cffi==23.1.0
bcrypt==4.2.0
PyJWT[crypto]==2.9.0

# --- Utilities ---
pydantic[email]==2.10.3
pydantic-settings==2.7.0
orjson==3.10.12
httpx==0.28.1
watchfiles==1.0.3

# --- Archive Handling ---
py7zr==0.22.0
rarfile==4.2

# --- CAD ---
ezdxf==1.3.4
